requirements.txt
ansi2html==1.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
docopt==0.6.2
Flask==2.2.5
fonttools==4.42.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
influxdb-client==1.37.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
nest-asyncio==1.5.7
numpy==1.23.5
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.16.1
pyepw==0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reactivex==4.0.4
requests==2.31.0
retrying==1.3.4
scipy==1.11.2
shapely==2.0.1
six==1.16.0
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
tzwhere==3.0.3
urllib3==1.26.16
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.16.2
